requirement.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
paho-mqtt==1.6.1
peewee==3.17.0
python-dotenv==1.0.0
simple-websocket==1.0.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
